# backend/src/hostings/hostings.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  HttpStatus,
  Param,
  Patch,
  Post,
  UseGuards,
} from "@nestjs/common";
import {
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
} from "@nestjs/swagger";
import { Permission } from "@prisma/client";
import { RequiredPermissions } from "src/common/decorators/required-permissions.decorator";
import { PermissionGuard } from "src/common/guards/permission.guard";
import { UserId } from "../common/decorators/user-id.decorator";
import {
  CreateHostingDto,
  HostingDto,
  UpdateHostingDto,
} from "./dto/hosting.dto";
import { HostingsService } from "./hostings.service";

@ApiTags("Hostings")
@Controller("hostings")
export class HostingsController {
  constructor(private readonly hostingService: HostingsService) {}

  @Get("count")
  @ApiOperation({
    summary: "Récupérer le nombre total d'hébergements (toutes applications)",
  })
  @ApiOkResponse({
    description: "Nombre total d'hébergements",
    type: Number,
  })
  public async countAllHostings(): Promise<number> {
    return this.hostingService.count();
  }
}

@ApiTags("Hostings")
@UseGuards(PermissionGuard)
@Controller("applications/:applicationId/hostings")
@ApiParam({
  name: "applicationId",
  description: "ID de l'application",
  type: String,
})
export class ApplicationHostingsController {
  constructor(private readonly hostingService: HostingsService) {}

  @Post()
  @RequiredPermissions([Permission.writeHostings])
  @ApiOperation({ summary: "Créer un hébergement pour une application" })
  @ApiCreatedResponse({
    description: "Hébergement créé",
    type: HostingDto,
  })
  @HttpCode(HttpStatus.CREATED)
  create(
    @UserId() userId: string,
    @Param("applicationId") applicationId: string,
    @Body() dto: CreateHostingDto,
  ) {
    // Ajoute l'ID de l'application provenant de l'URL dans le DTO
    return this.hostingService.createHosting({ ...dto, applicationId }, userId);
  }

  @Get()
  @RequiredPermissions([Permission.readHostings])
  @ApiOperation({
    summary: "Récupérer tous les hébergements d'une application",
  })
  @ApiOkResponse({
    description: "Liste des hébergements",
    type: HostingDto,
    isArray: true,
  })
  findAll(@Param("applicationId") applicationId: string) {
    return this.hostingService.findByApplicationId(applicationId);
  }

  @Get(":id")
  @RequiredPermissions([Permission.readHostings])
  @ApiOperation({
    summary: "Récupérer un hébergement par ID pour une application",
  })
  @ApiOkResponse({
    description: "Hébergement trouvé",
    type: HostingDto,
  })
  findOne(@Param("id") id: string) {
    return this.hostingService.findOneHosting(id);
  }

  @Patch(":id")
  @RequiredPermissions([Permission.writeHostings])
  @ApiOperation({
    summary: "Mettre à jour un hébergement pour une application",
  })
  @ApiOkResponse({
    description: "Hébergement mis à jour",
    type: HostingDto,
  })
  update(
    @UserId() userId: string,
    @Param("id") id: string,
    @Param("applicationId") applicationId: string,
    @Body() dto: UpdateHostingDto,
  ) {
    return this.hostingService.updateHosting(
      id,
      { ...dto, applicationId },
      userId,
    );
  }

  @Delete(":id")
  @RequiredPermissions([Permission.writeHostings])
  @ApiOperation({ summary: "Supprimer un hébergement pour une application" })
  @ApiNoContentResponse({ description: "Hébergement supprimé" })
  @HttpCode(HttpStatus.NO_CONTENT)
  async remove(@UserId() userId: string, @Param("id") id: string) {
    await this.hostingService.remove(id, userId);
  }
}