# frontend/openapi/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/v2/business-division:
    get:
      operationId: BusinessDivisionController_findAll
      summary: Rechercher des Business division.
      description: "

        \    Cette route permet de rechercher selon les paramètres fournis :

        \ \ 

        \    - Retourne une liste paginée ou non de business division.

        \    - Filtrage possible par nom.

        \    - Tri possible par :

        \      - `label`

        \    - Ordre : `asc` ou `desc`.

        \  "
      parameters:
        - name: page
          required: false
          in: query
          description: Numéro de page
          schema:
            minimum: 0
            example: 0
            type: number
        - name: pageSize
          required: false
          in: query
          description: Nombre de résultats par page, 0 pour supprimer la pagination
          schema:
            minimum: 0
            default: 15
            example: 15
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ utilisé pour le tri
          schema:
            example: label
            type: string
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: asc
            enum:
              - asc
              - desc
            type: string
        - name: label
          required: false
          in: query
          description: Recherche sur le nom
          schema:
            example: farm
            type: string
      responses:
        "200":
          description: Liste des tags trouvés
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBusinessDivisionDTO"
      tags: &a8
        - Business Division
  /api/v2/business-division/{id}:
    get:
      operationId: BusinessDivisionController_findById
      summary: Rechercher un Business division par id.
      parameters:
        - name: id
          required: true
          in: path
          description: ID du Business Division à trouver
          schema:
            type: string
      responses:
        "200":
          description: Business division
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDivisionDTO"
        "404":
          description: Business Division non trouvé
      tags: *a8
  /api/v2/applications/{applicationId}/hostings:
    post:
      operationId: ApplicationHostingsController_create
      summary: Créer un hébergement pour une application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostingDto"
      responses:
        "201":
          description: Hébergement créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingDto"
      tags: &a9
        - Hostings
    get:
      operationId: ApplicationHostingsController_findAll
      summary: Récupérer tous les hébergements d'une application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      responses:
        "200":
          description: Liste des hébergements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostingDto"
      tags: *a9
  /api/v2/applications/{applicationId}/hostings/{id}:
    get:
      operationId: ApplicationHostingsController_findOne
      summary: Récupérer un hébergement par ID pour une application
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      responses:
        "200":
          description: Hébergement trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingDto"
      tags: *a9
    patch:
      operationId: ApplicationHostingsController_update
      summary: Mettre à jour un hébergement pour une application
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHostingDto"
      responses:
        "200":
          description: Hébergement mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingDto"
      tags: *a9
    delete:
      operationId: ApplicationHostingsController_remove
      summary: Supprimer un hébergement pour une application
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: applicationId
          required: true
          in: path
          description: ID de l'application
          schema:
            type: string
      responses:
        "204":
          description: Hébergement supprimé
      tags: *a9
...
  /api/v2/applications/{applicationId}/compliances:
    post:
      operationId: ApplicationCompliancesController_create
      summary: Create a new compliance for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplianceDto"
      responses:
        "201":
          description: Compliance created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceDto"
        "409":
          description: A compliance already exists for this application
      tags: &a20
        - Compliances
    get:
      operationId: ApplicationCompliancesController_findOne
      summary: Retrieve the compliance for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      responses:
        "200":
          description: Compliance found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceDto"
      tags: *a20
    patch:
      operationId: ApplicationCompliancesController_update
      summary: Update the compliance for an application
      parameters:
        - name: applicationId
          required: true
          in: path
          description: ID of the application
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceDto"
      responses:
        "200":
          description: Compliance updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceDto"
      tags: *a20
...
info:
  title: API Référentiel Applications
  description: API pour la gestion des applications
  version: "2.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/auth
          tokenUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          refreshUrl: http://localhost:8082/realms/referentiel-applications//protocol/openid-connect/token
          scopes:
            openid: OpenID scope
            profile: Profile scope
      description: OAuth2 authentication using OIDC
    api_key:
      type: apiKey
      in: header
      name: x-refapp-token
      description: Token authentication
  schemas:
...
    BusinessDivisionDTO:
      type: object
      properties:
        id:
          type: string
          description: Identifiant unique de la division DTO
        label:
          type: string
          description: Nom unique du business division
          example: farm
      required:
        - id
        - label
...
    PaginatedBusinessDivisionDTO:
      type: object
      properties:
        results:
          description: Array of results
          type: array
          items:
            $ref: "#/components/schemas/BusinessDivisionDTO"
        total:
          type: number
          description: Total number of items
          example: 100
      required:
        - results
        - total
    CreateHostingDto:
      type: object
      properties:
        label:
          type: string
        hostingOptionId:
          type: string
          description: ID of the hosting option
        applicationId:
          type: string
        isActive:
          type: boolean
          nullable: true
          description: Indicates whether the hosting is active, passif or not specified
            (null).
      required:
        - applicationId
        - isActive
    HostingOptionDto:
      type: object
      properties:
        site:
          type: string
          description: Site name (géographique)
          example: CER(RENNES)
        platform:
          type: string
          description: Platform name
          example: PHYSIQUE
        provider:
          type: string
          description: Provider name
          example: DTNUM
        building:
          type: string
          description: Building (optional)
          example: B15
        room:
          type: string
          description: Room (optional)
          example: IT2
        id:
          type: string
          description: Unique identifier of the hosting option
          example: 12345678-1234-1234-1234-123456789012
      required:
        - site
        - platform
        - provider
        - id
    HostingDto:
      type: object
      properties:
        label:
          type: string
        hostingOptionId:
          type: string
          description: ID of the hosting option
        applicationId:
          type: string
        isActive:
          type: boolean
          nullable: true
          description: Indicates whether the hosting is active, passif or not specified
            (null).
        id:
          type: string
        hostingOption:
          $ref: "#/components/schemas/HostingOptionDto"
      required:
        - applicationId
        - isActive
        - id
        - hostingOption
    UpdateHostingDto:
      type: object
      properties:
        label:
          type: string
        hostingOptionId:
          type: string
          description: ID of the hosting option
        applicationId:
          type: string
        isActive:
          type: boolean
          nullable: true
          description: Indicates whether the hosting is active, passif or not specified
            (null).
...
    CreateComplianceDto:
      type: object
      properties:
        dima_duration_hours:
          type: number
          example: 4
          description: DIMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        dima_is_hno:
          type: boolean
          example: true
          description: DIMA HNO (Heure non ouvrée)
        dima_business_impact:
          type: string
          example: Perte de revenu, impact client
          description: DIMA business impact of interruption
        dima_recovery_plan:
          type: boolean
          example: true
          description: DIMA recovery plan exists
        dima_recovery_solutions:
          type: string
          example: Bascule sur serveur de secours
          description: DIMA recovery solutions
        dima_last_test_date:
          type: string
          example: 2023-01-01
          description: DIMA last test date
        dima_test_result:
          enum: &a27
            - OK
            - KO
          type: string
          description: DIMA test result
        dima_recovery_manager:
          type: string
          example: Jean Dupont
          description: DIMA recovery manager
        pdma_duration_hours:
          type: number
          example: 8
          description: PDMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        pdma_data_types:
          type: string
          example: Transactions, logs, fichiers utilisateur
          description: PDMA data types concerned
        pdma_backup_frequency:
          type: string
          example: Toutes les heures
          description: PDMA backup frequency
        pdma_backup_method:
          type: string
          example: Snapshot, backup incrémental
          description: PDMA backup method used
        pdma_backup_storage:
          enum: &a28
            - S3
            - LOCAL
            - EXTERNE
          type: string
          description: PDMA backup storage
        pdma_last_test_date:
          type: string
          example: 2023-01-01
          description: PDMA last test date
        pdma_test_result:
          enum: &a29
            - OK
            - KO
          type: string
          description: PDMA test result
        pdma_restoration_manager:
          type: string
          example: Marie Martin
          description: PDMA restoration manager
        homologation_status:
          enum: &a30
            - homologuee
            - en_cours
            - dispensee
          type: string
          description: Homologation status
        homologation_date_end:
          type: string
          example: 2023-01-01
          description: Homologation date end
        homologation_rssi_id:
          type: string
          example: uuid-of-rssi-actor
          description: Homologation RSSI actor ID
        rgaa_audit_date:
          type: string
          example: 2023-01-01
          description: RGAA audit date
        rgaa_service_url:
          type: string
          example: https://service.example.com
          description: RGAA service URL
        rgaa_accessibility_url:
          type: string
          example: https://service.example.com/accessibilite
          description: RGAA accessibility page URL
        rgaa_score_percentage:
          type: string
          minimum: 0
          maximum: 100
          example: 85
          description: RGAA score percentage (0-100)
        dsfr_implemented:
          type: boolean
          example: true
          description: DSFR implemented
        dsfr_version:
          type: string
          example: 1.9.0
          description: DSFR version used
        rgpd_has_aipd:
          type: boolean
          example: true
          description: RGPD has AIPD
        rgpd_dpo_name:
          type: string
          example: Pierre Durand
          description: RGPD DPO name
    ComplianceDto:
      type: object
      properties:
        dima_duration_hours:
          type: number
          example: 4
          description: DIMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        dima_is_hno:
          type: boolean
          example: true
          description: DIMA HNO (Heure non ouvrée)
        dima_business_impact:
          type: string
          example: Perte de revenu, impact client
          description: DIMA business impact of interruption
        dima_recovery_plan:
          type: boolean
          example: true
          description: DIMA recovery plan exists
        dima_recovery_solutions:
          type: string
          example: Bascule sur serveur de secours
          description: DIMA recovery solutions
        dima_last_test_date:
          type: string
          example: 2023-01-01
          description: DIMA last test date
        dima_test_result:
          enum: *a27
          type: string
          description: DIMA test result
        dima_recovery_manager:
          type: string
          example: Jean Dupont
          description: DIMA recovery manager
        pdma_duration_hours:
          type: number
          example: 8
          description: PDMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        pdma_data_types:
          type: string
          example: Transactions, logs, fichiers utilisateur
          description: PDMA data types concerned
        pdma_backup_frequency:
          type: string
          example: Toutes les heures
          description: PDMA backup frequency
        pdma_backup_method:
          type: string
          example: Snapshot, backup incrémental
          description: PDMA backup method used
        pdma_backup_storage:
          enum: *a28
          type: string
          description: PDMA backup storage
        pdma_last_test_date:
          type: string
          example: 2023-01-01
          description: PDMA last test date
        pdma_test_result:
          enum: *a29
          type: string
          description: PDMA test result
        pdma_restoration_manager:
          type: string
          example: Marie Martin
          description: PDMA restoration manager
        homologation_status:
          enum: *a30
          type: string
          description: Homologation status
        homologation_date_end:
          type: string
          example: 2023-01-01
          description: Homologation date end
        homologation_rssi_id:
          type: string
          example: uuid-of-rssi-actor
          description: Homologation RSSI actor ID
        rgaa_audit_date:
          type: string
          example: 2023-01-01
          description: RGAA audit date
        rgaa_service_url:
          type: string
          example: https://service.example.com
          description: RGAA service URL
        rgaa_accessibility_url:
          type: string
          example: https://service.example.com/accessibilite
          description: RGAA accessibility page URL
        rgaa_score_percentage:
          type: string
          minimum: 0
          maximum: 100
          example: 85
          description: RGAA score percentage (0-100)
        dsfr_implemented:
          type: boolean
          example: true
          description: DSFR implemented
        dsfr_version:
          type: string
          example: 1.9.0
          description: DSFR version used
        rgpd_has_aipd:
          type: boolean
          example: true
          description: RGPD has AIPD
        rgpd_dpo_name:
          type: string
          example: Pierre Durand
          description: RGPD DPO name
        id:
          type: string
          example: 12345678-1234-1234-1234-123456789012
          description: Unique identifier of the compliance record
      required:
        - id
    UpdateComplianceDto:
      type: object
      properties:
        dima_duration_hours:
          type: number
          example: 4
          description: DIMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        dima_is_hno:
          type: boolean
          example: true
          description: DIMA HNO (Heure non ouvrée)
        dima_business_impact:
          type: string
          example: Perte de revenu, impact client
          description: DIMA business impact of interruption
        dima_recovery_plan:
          type: boolean
          example: true
          description: DIMA recovery plan exists
        dima_recovery_solutions:
          type: string
          example: Bascule sur serveur de secours
          description: DIMA recovery solutions
        dima_last_test_date:
          type: string
          example: 2023-01-01
          description: DIMA last test date
        dima_test_result:
          enum: *a27
          type: string
          description: DIMA test result
        dima_recovery_manager:
          type: string
          example: Jean Dupont
          description: DIMA recovery manager
        pdma_duration_hours:
          type: number
          example: 8
          description: PDMA duration in hours (1, 4, 8, 12, 24, 48, 72)
        pdma_data_types:
          type: string
          example: Transactions, logs, fichiers utilisateur
          description: PDMA data types concerned
        pdma_backup_frequency:
          type: string
          example: Toutes les heures
          description: PDMA backup frequency
        pdma_backup_method:
          type: string
          example: Snapshot, backup incrémental
          description: PDMA backup method used
        pdma_backup_storage:
          enum: *a28
          type: string
          description: PDMA backup storage
        pdma_last_test_date:
          type: string
          example: 2023-01-01
          description: PDMA last test date
        pdma_test_result:
          enum: *a29
          type: string
          description: PDMA test result
        pdma_restoration_manager:
          type: string
          example: Marie Martin
          description: PDMA restoration manager
        homologation_status:
          enum: *a30
          type: string
          description: Homologation status
        homologation_date_end:
          type: string
          example: 2023-01-01
          description: Homologation date end
        homologation_rssi_id:
          type: string
          example: uuid-of-rssi-actor
          description: Homologation RSSI actor ID
        rgaa_audit_date:
          type: string
          example: 2023-01-01
          description: RGAA audit date
        rgaa_service_url:
          type: string
          example: https://service.example.com
          description: RGAA service URL
        rgaa_accessibility_url:
          type: string
          example: https://service.example.com/accessibilite
          description: RGAA accessibility page URL
        rgaa_score_percentage:
          type: string
          minimum: 0
          maximum: 100
          example: 85
          description: RGAA score percentage (0-100)
        dsfr_implemented:
          type: boolean
          example: true
          description: DSFR implemented
        dsfr_version:
          type: string
          example: 1.9.0
          description: DSFR version used
        rgpd_has_aipd:
          type: boolean
          example: true
          description: RGPD has AIPD
        rgpd_dpo_name:
          type: string
          example: Pierre Durand
          description: RGPD DPO name
...
security:
  - oauth2: []
  - api_key: []